Feature: Set up DB

+++ db/migrate/20220609164315_create_teas.rb
@@ -3,7 +3,7 @@ class CreateTeas < ActiveRecord::Migration[5.2]
     create_table :teas do |t|
       t.string :title
       t.text :description
-      t.float :temperature
+      t.integer :temperature
       t.integer :brew_time
 
       t.timestamps
